+++ src/Settings/Display/DisplaySettings.ts
@@ -1,7 +1,6 @@
 import { makeAutoObservable } from 'mobx';
 import { z } from 'zod';
 
-import { App } from 'src/App';
 import { ApplyableNumberField } from 'src/Components/ApplyableTextField';
 
 /** Enumerates the different possible ways the TX and RX data
@@ -21,8 +20,6 @@ export const dataViewConfigEnumToDisplayName: {
 };
 export default class DisplaySettings {
 
-  app: App;
-
   terminalWidthChars = new ApplyableNumberField('120', z.coerce.number().int().min(1));
 
   scrollbackBufferSizeRows = new ApplyableNumberField('2000', z.coerce.number().int().min(1));
@@ -32,8 +29,7 @@ export default class DisplaySettings {
   // 14px is a good default size for the terminal text
   charSizePx = new ApplyableNumberField('14', z.coerce.number().int().min(1));
 
-  constructor(app: App) {
-    this.app = app;
+  constructor() {
     makeAutoObservable(this);
   }