+++ src/Terminals/SingleTerminal/SingleTerminalRow.tsx
@@ -13,4 +13,15 @@ export default class TerminalRow {
     this.terminalChars = [];
     makeAutoObservable(this);
   }
+
+  /**
+   * Returns the raw text of the row, by joining all the individual chars together
+   * into a single string. Discards all other properties of a terminal char, such
+   * as formatting.
+   *
+   * @returns The raw text of the row.
+   */
+  getText(): string {
+    return this.terminalChars.map((terminalChar) => terminalChar.char).join('');
+  }
 }