+++ src/Terminals/Terminals.ts
@@ -16,9 +16,9 @@ export default class Terminals {
   filterText: ApplyableTextField
 
   constructor(app: App) {
-    this.txRxTerminal = new Terminal(app, true);
-    this.rxTerminal = new Terminal(app, false); // Not focusable
-    this.txTerminal = new Terminal(app, true);
+    this.txRxTerminal = new Terminal(true, app.settings.dataProcessingSettings, app.settings.displaySettings, app.handleTerminalKeyDown);
+    this.rxTerminal = new Terminal(false, app.settings.dataProcessingSettings, app.settings.displaySettings, app.handleTerminalKeyDown); // Not focusable
+    this.txTerminal = new Terminal(true, app.settings.dataProcessingSettings, app.settings.displaySettings, app.handleTerminalKeyDown);
 
     this.filterText = new ApplyableTextField('', z.string());
     this.filterText.setOnApplyChanged(this.onFilterTextApply);